chore: Split `CameraSession` into multiple files
