+++ package/android/src/main/java/com/mrousavy/camera/core/utils/FileUtils.kt
@@ -2,6 +2,8 @@ package com.mrousavy.camera.core.utils
 
 import android.content.Context
 import android.graphics.Bitmap
+import android.graphics.BitmapFactory
+import android.util.Size
 import java.io.File
 import java.io.FileOutputStream
 
@@ -16,6 +18,16 @@ class FileUtils {
       FileOutputStream(file).use { stream ->
         bitmap.compress(Bitmap.CompressFormat.JPEG, quality, stream)
       }
+    }
+
+    fun getImageSize(imagePath: String): Size {
+      val bitmapOptions = BitmapFactory.Options().also {
+        it.inJustDecodeBounds = true
+      }
+      BitmapFactory.decodeFile(imagePath, bitmapOptions)
+      val width = bitmapOptions.outWidth
+      val height = bitmapOptions.outHeight
+      return Size(width, height)
     }
   }
 }
